+++ lib/Model/Queue/ExternalDistanceLimitingQueue.php
@@ -4,7 +4,6 @@ namespace DTL\Extension\Fink\Model\Queue;
 
 use DTL\Extension\Fink\Model\Url;
 use DTL\Extension\Fink\Model\UrlQueue;
-use DTL\Extension\Fink\Model\Urls;
 
 class ExternalDistanceLimitingQueue implements UrlQueue
 {