+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Support for custom headers via. `--header=Foo:Bar` #46
 - Support multiple base URLs #44
+- Support setting the `--rate` (max number of requests per second) #51
 
 ## [0.4.1]