+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [0.5.0]
 
 - Support for custom headers via. `--header=Foo:Bar` #46
 - Support multiple base URLs #44